added casts to fix polars bug

+++ src/PCN_library.py
@@ -1181,11 +1181,12 @@ def naive_themisto_PCN_estimation(themisto_results_csv_file, replicon_length_csv
     ##Also note that this function omits replicons with zero mapped reads.
     print("running naive themisto PCN estimation (ignoring multireplicon reads)")
     ## import the data as polars dataframes.
+    
     replicon_length_df = pl.read_csv(replicon_length_csv_file)
     naive_themisto_read_count_df = pl.read_csv(themisto_results_csv_file).filter(
         (pl.col("SeqType") == "chromosome") | (pl.col("SeqType") == "plasmid")).join(
             replicon_length_df, on = "SeqID").with_columns(
-                (pl.col("ReadCount") / pl.col("replicon_length")).alias("SequencingCoverage"))
+                (pl.col("ReadCount").cast(pl.Float64) / pl.col("replicon_length").cast(pl.Float64)).alias("SequencingCoverage"))
     
     ## make a second dataframe containing just the sequencing coverage for the longest replicon for each genome.
     ## to do so, first group by AnnotationAccession and compute maximum replicon_length within each group.
@@ -1603,7 +1604,7 @@ def make_PIRAGenomeDataFrame(
                         ## sum those ReadCounts,
                         (pl.col("InitialReadCount") + pl.col("AdditionalReadCount")).alias("ReadCount")).with_columns(
                             ## and re-calculate SequencingCoverage,
-                            (pl.col("ReadCount") / pl.col("replicon_length")).alias("SequencingCoverage"))
+                            (pl.col("ReadCount").cast(pl.Float64) / pl.col("replicon_length").cast(pl.Float64)).alias("SequencingCoverage"))
     
     ## The following is a hack, following code in the function native_themisto_PCN_estimation(),
     ## to recalculate LongestRepliconCoverage and CopyNumber with the additional reads.
